django upgrade

+++ freppledb/reportmanager/models.py
@@ -52,7 +52,9 @@ class SQLReport(AuditModel):
             db = getattr(req, "database", DEFAULT_DB_ALIAS)
         else:
             db = getattr(_thread_locals, "database", DEFAULT_DB_ALIAS)
-        SQLColumn.objects.filter(report=self).using(db).delete()
+        if self.pk:
+            SQLColumn.objects.filter(report=self).using(db).delete()
+
         if self.sql:
 
             with connections[
@@ -61,29 +63,37 @@ class SQLReport(AuditModel):
 
                 # The query is limited to 1 row. We just need the column names at this point.
                 cursor.execute("select * from (%s) as Q limit 1" % self.sql)
-                if self.id:
-                    cols = []
-                    seq = 1
+                # For unsaved instances we only validate SQL syntax and column names.
+                if not self.pk:
+                    cols = set()
                     for f in cursor.description:
                         if f[0] in cols:
                             raise Exception("Duplicate column name '%s'" % f[0])
-                        cols.append(f[0])
-                        if f[1] == 1700:
-                            fmt = "number"
-                        elif f[1] == 1184:
-                            fmt = "datetime"
-                        elif f[1] in (23, 20):
-                            fmt = "integer"
-                        elif f[1] == 1186:
-                            fmt = "duration"
-                        elif f[1] == 1043:
-                            fmt = "text"
-                        else:
-                            fmt = "character"
-                        SQLColumn(
-                            report=self, sequence=seq, name=f[0], format=fmt
-                        ).save(using=db)
-                        seq += 1
+                        cols.add(f[0])
+                    return
+
+                cols = []
+                seq = 1
+                for f in cursor.description:
+                    if f[0] in cols:
+                        raise Exception("Duplicate column name '%s'" % f[0])
+                    cols.append(f[0])
+                    if f[1] == 1700:
+                        fmt = "number"
+                    elif f[1] == 1184:
+                        fmt = "datetime"
+                    elif f[1] in (23, 20):
+                        fmt = "integer"
+                    elif f[1] == 1186:
+                        fmt = "duration"
+                    elif f[1] == 1043:
+                        fmt = "text"
+                    else:
+                        fmt = "character"
+                    SQLColumn(report=self, sequence=seq, name=f[0], format=fmt).save(
+                        using=db
+                    )
+                    seq += 1
 
     def save(self, *args, **kwargs):
         if self.sql and "pg_" in self.sql.lower():
